✂️  Remove Postman page
The goal of this project is to document Procore API and how to use it.

There are several pages in which we make reference to Postman and link to
its website and snippets, for instance:

- https://github.com/procore/documentation/blob/main/additional_resources/faq.md#testing
- https://github.com/procore/documentation/blob/main/getting_started/making_first_call.md#curl-and-postman
- https://github.com/procore/documentation/blob/main/oauth/oauth_introduction.md#procore-oauth-20-postman-collection-with-curl-examples

These references are helpful enough for developers to check Postman if they
are interested. However, having an entire page about Postman itself goes
beyond the scope of this project.

By removing a page that is not the *focus* of this project, we make it easier
for developers to find the specific help they are looking for in these guides.